End screen, new fonts, loading screen, visuals
Made cells look nicer, moved colour chart to right of grid to prepare
for UI at bottom.

Added icon that appears when a cell has only one possible choice (is
locked) to make the game easier.

Added display that shows the currently available colours for the cell
that is being hovered over.

Reworked the end game screen and fixed a crashing bug in it.